+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for copying objects within OSS bucket
+
 ## [5.0.0] - 2019-09-03
 
 ### Changed